+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Fixed goroutine leak on close reader
 * Fixed topic reader and writer WaitInit hunging on unretriable connection error
 * Added `query.Client.Stats()` method
 * Added `query.Result.Stats()` method